requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
starlette>=0.27.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6
python-dotenv>=1.0.0

# Azure SDK
azure-storage-blob>=12.16.0
azure-identity>=1.13.0
azure-eventhub>=5.11.2
azure-data-tables>=12.4.2
azure-cosmos>=4.3.1

# Database and data processing
sqlalchemy>=2.0.0
asyncpg>=0.27.0
pandas>=2.0.0
pyarrow>=13.0.0
openpyxl>=3.1.0

# Monitoring
prometheus-client>=0.17.0
psutil>=5.9.0

# HTTP and async
aiohttp>=3.8.4
httpx>=0.24.1

# Testing
pytest>=7.3.1
pytest-asyncio>=0.21.0